agent-framework/first-agent/pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework-ollama>=1.0.0b260423",
]
